refactoring and style improvments
