+++ src/utils/colors.js
@@ -1,0 +1,2 @@
+export const primaryColor = "rgb(37, 99, 235)"
+export const dangerColor = "rgb(220, 38, 38)"
